# ASPNETCore5Sample/Models/Course.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
using Newtonsoft.Json;
using System;
using System.Collections.Generic;

namespace ASPNETCore5Sample.Models
{
    public partial class Course
    {
        public Course()
        {
            CourseInstructor = new HashSet<CourseInstructor>();
            Enrollment = new HashSet<Enrollment>();
        }

        public int CourseId { get; set; }
        public string Title { get; set; }
        public int Credits { get; set; }
        public int DepartmentId { get; set; }
        public DateTime? DateModified { get; set; }
        public bool IsDeleted { get; set; }

        [JsonIgnore]
        public virtual Department Department { get; set; }
        [JsonIgnore]
        public virtual ICollection<CourseInstructor> CourseInstructor { get; set; }
        [JsonIgnore]
        public virtual ICollection<Enrollment> Enrollment { get; set; }
    }
}