# ASPNETCore5Sample/Models/OfficeAssignment.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
using Newtonsoft.Json;

namespace ASPNETCore5Sample.Models
{
    public partial class OfficeAssignment
    {
        public int InstructorId { get; set; }
        public string Location { get; set; }

        [JsonIgnore]
        public virtual Person Instructor { get; set; }
    }
}