# ASPNETCore5Sample/Models/Person.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
using Newtonsoft.Json;
using System;
using System.Collections.Generic;

namespace ASPNETCore5Sample.Models
{
    public partial class Person
    {
        public Person()
        {
            CourseInstructor = new HashSet<CourseInstructor>();
            Department = new HashSet<Department>();
            Enrollment = new HashSet<Enrollment>();
        }

        public int Id { get; set; }
        public string LastName { get; set; }
        public string FirstName { get; set; }
        public DateTime? HireDate { get; set; }
        public DateTime? EnrollmentDate { get; set; }
        public string Discriminator { get; set; }
        public DateTime? DateModified { get; set; }
        public bool IsDeleted { get; set; }

        [JsonIgnore]
        public virtual OfficeAssignment OfficeAssignment { get; set; }
        [JsonIgnore]
        public virtual ICollection<CourseInstructor> CourseInstructor { get; set; }
        [JsonIgnore]
        public virtual ICollection<Department> Department { get; set; }
        [JsonIgnore]
        public virtual ICollection<Enrollment> Enrollment { get; set; }
    }
}